# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Node MySQL API
  version: 1.0.0
  description: Express, MySQL, Sequelize Boilerplate API with JWT & Refresh Token Authentication
servers:
  - url: http://localhost:4000
paths:
  /accounts/authenticate:
    post:
      summary: Authenticate credentials and return a JWT token and cookie-based refresh token
      tags:
        - Accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthenticateRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticateResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=abc123xyz; Path=/; HttpOnly; Expires=...
        '400':
          description: Invalid email or password

  /accounts/refresh-token:
    post:
      summary: Use a refresh token to generate a new JWT token and rotate the refresh token
      tags:
        - Accounts
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticateResponse'
        '400':
          description: Invalid or expired refresh token

  /accounts/revoke-token:
    post:
      summary: Revoke a refresh token so it can no longer be used
      tags:
        - Accounts
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: Optional refresh token to revoke. If omitted, the cookie token is revoked.
      responses:
        '200':
          description: Token revoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Token revoked
        '400':
          description: Token is required
        '401':
          description: Unauthorized

  /accounts/register:
    post:
      summary: Register a new user account and send verification email
      tags:
        - Accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Registration successful, please check your email for verification instructions

  /accounts/verify-email:
    post:
      summary: Verify a new account using the token sent via email
      tags:
        - Accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Verification successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Verification successful, you can now login
        '400':
          description: Verification failed

  /accounts/forgot-password:
    post:
      summary: Request a password reset token sent via email
      tags:
        - Accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset request received
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Please check your email for password reset instructions

  /accounts/validate-reset-token:
    post:
      summary: Validate a password reset token before displaying the reset page
      tags:
        - Accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateResetTokenRequest'
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Token is valid
        '400':
          description: Invalid token

  /accounts/reset-password:
    post:
      summary: Reset account password using the token
      tags:
        - Accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password reset successful, you can now login
        '400':
          description: Validation error

  /accounts:
    get:
      summary: Get all accounts (Admin only)
      tags:
        - Accounts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    post:
      summary: Create a new account directly (Admin only)
      tags:
        - Accounts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '200':
          description: Account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  firstName:
                    type: string
                  lastName:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
                  created:
                    type: string
                  updated:
                    type: string
                  isVerified:
                    type: boolean
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  /accounts/me:
    get:
      summary: Get details of the currently authenticated account
      tags:
        - Accounts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          description: Unauthorized

  /accounts/{id}:
    get:
      summary: Get account by ID (Self or Admin)
      tags:
        - Accounts
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the account
          schema:
            type: integer
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          description: Unauthorized
        '404':
          description: Account not found
    put:
      summary: Update account by ID (Self or Admin)
      tags:
        - Accounts
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the account to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountRequest'
      responses:
        '200':
          description: Account updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          description: Unauthorized
        '404':
          description: Account not found
    delete:
      summary: Delete an account (Self or Admin)
      tags:
        - Accounts
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the account to delete
          schema:
            type: integer
      responses:
        '200':
          description: Account deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Account deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Account not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Account:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Mr
        firstName:
          type: string
          example: System
        lastName:
          type: string
          example: Admin
        email:
          type: string
          example: admin@example.com
        role:
          type: string
          example: admin
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
          nullable: true
        isVerified:
          type: boolean
          example: true
    AuthenticateRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: admin@example.com
        password:
          type: string
          format: password
          example: Password123!
    AuthenticateResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Mr
        firstName:
          type: string
          example: System
        lastName:
          type: string
          example: Admin
        email:
          type: string
          example: admin@example.com
        role:
          type: string
          example: admin
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
          nullable: true
        isVerified:
          type: boolean
          example: true
        jwtToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          example: a4e830cd4416f4baa4acb6876292a21d...
    RegisterRequest:
      type: object
      required:
        - title
        - firstName
        - lastName
        - email
        - password
        - confirmPassword
        - acceptTerms
      properties:
        title:
          type: string
          example: Mr
        firstName:
          type: string
          example: Alice
        lastName:
          type: string
          example: Smith
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          format: password
          minLength: 6
          example: password123
        confirmPassword:
          type: string
          format: password
          example: password123
        acceptTerms:
          type: boolean
          enum: [true]
          example: true
    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          example: b855a419d99696f65925281c7b7cf02a238b4bee...
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
    ValidateResetTokenRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          example: 4c3e80c4416f4baa4acb6876292a21d042025cb...
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
        - confirmPassword
      properties:
        token:
          type: string
          example: 4c3e80c4416f4baa4acb6876292a21d042025cb...
        password:
          type: string
          format: password
          minLength: 6
          example: newpassword123
        confirmPassword:
          type: string
          format: password
          example: newpassword123
    CreateAccountRequest:
      type: object
      required:
        - title
        - firstName
        - lastName
        - email
        - password
        - confirmPassword
        - role
      properties:
        title:
          type: string
          example: Mr
        firstName:
          type: string
          example: Bob
        lastName:
          type: string
          example: Jones
        email:
          type: string
          format: email
          example: bob@example.com
        password:
          type: string
          format: password
          minLength: 6
          example: password123
        confirmPassword:
          type: string
          format: password
          example: password123
        role:
          type: string
          enum: [Admin, User]
          example: User
    UpdateAccountRequest:
      type: object
      properties:
        title:
          type: string
          example: Mrs
        firstName:
          type: string
          example: Alice
        lastName:
          type: string
          example: Jones
        email:
          type: string
          format: email
          example: alice.jones@example.com
        password:
          type: string
          format: password
          minLength: 6
          example: newsecurepass123
        confirmPassword:
          type: string
          format: password
          example: newsecurepass123
        role:
          type: string
          enum: [Admin, User]
          example: User